# api/api.yaml
openapi: 3.0.3
info:
  title: Emote Tracker API
  description: The Rest API for the Emote Tracker
  version: 1.0.0
paths:
  /user/register:
    get:
      tags:
        - registration
      summary: Register a new user to track the emote usage
      operationId: registerNewUser
      parameters:
        - in: query
          name: userId
          required: true
          description: Twitch User ID
          schema:
            type: string
      responses:
        200:
          description: Successfully registered new user
          content:
            'application/json':
              schema:
                type: string
        304:
          description: Nothing was changed. User already registered
          content:
            'application/json':
              schema:
                type: string

  /user/unregister:
    get:
      tags:
        - registration
      operationId: unregisterUser
      parameters:
        - in: query
          name: userId
          required: true
          description: Twitch User ID
          schema:
            type: string
      responses:
        200:
          description: Successfully unregistered user
          content:
            'application/json':
              schema:
                type: string
        304:
          description: Nothing was changed. User already unregistered/ never registered
          content:
            'application/json':
              schema:
                type: string

  /user/overview:
    get:
      tags:
        - user
      operationId: getUserOverview
      parameters:
        - in: query
          name: userId
          schema:
            type: string
      responses:
        200:
          description: Successfully got user overview
          content:
            'application/json':
              schema:
                type: string

  /user/{userId}/emotes/count/top:
    get:
      tags:
        - user
      operationId: getTopEmotes
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: count
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Successfully retrieved top emotes based on their usage count
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Emote'

  /user/{userId}/emotes/count/bottom:
    get:
      tags:
        - user
      operationId: getBottomEmotes
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: count
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Successfully retrieved bottom emotes based on their usage count
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Emote'

  /user/{userId}/emotes/count/above-average:
    get:
      tags:
        - user
      operationId: getAboveAverageEmotes
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successfully retrieved bottom emotes based on their usage count
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Emote'

  /user/{userId}/emotes/count/below-average:
    get:
      tags:
        - user
      operationId: getBelowAverageEmotes
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successfully retrieved bottom emotes based on their usage count
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Emote'

  /user/{userId}/emotes/count/no-usage:
    get:
      tags:
        - user
      operationId: getEmotesWitNoUsage
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get all emotes without usage
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Emote'

  /user/{userId}/emotes/update:
    patch:
      tags:
        - user
      operationId: updateEmotesForUser
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Updated emote set successfully

components:
  schemas:
    Emote:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        count:
          type: integer
          format: int32